requirements.txt
absl-py==0.12.0
antlr4-python3-runtime==4.9.2
anykeystore==0.2
appdirs==1.4.4
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cityscapesScripts==2.2.0
click==7.1.2
coloredlogs==15.0
configparser==5.0.2
cycler==0.10.0
decorator==5.0.7
defusedxml==0.7.1
docker-pycreds==0.4.0
entrypoints==0.3
flatbuffers==1.12
focal-loss-torch==0.0.7
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.0.0
grpcio==1.37.0
h5py==3.2.1
humanfriendly==9.1
hupper==1.10.2
idna==2.10
imageio==2.9.0
importlib-resources==5.1.2
ipykernel==5.5.3
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.20.2
oauthlib==3.1.0
omegaconf==2.0.6
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
pandocfilters==1.4.3
parso==0.8.2
PasteDeploy==2.1.1
pathtools==0.1.2
pbkdf2==1.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
plaster==1.0
plaster-pastedeploy==0.7
prometheus-client==0.10.1
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.15.8
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.7
pyquaternion==0.9.9
pyramid==2.0
pyramid-mailer==0.15.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytorch-msssim==0.2.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
q==2.6
repoze.sendmail==4.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.3
Send2Trash==1.5.0
sentry-sdk==1.0.0
shortuuid==1.0.1
six==1.15.0
smmap==4.0.0
SQLAlchemy==1.4.11
subprocess32==3.5.4
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.4.8
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
transaction==3.0.1
translationstring==1.4
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
velruse==1.1.1
venusian==3.0.0
wandb==0.10.28
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.7
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
WTForms==2.3.3
wtforms-recaptcha==0.3.2
zope.deprecation==4.4.0
zope.interface==5.4.0
zope.sqlalchemy==1.4
